Move ConcertProjectionRepository into example.application.projection package
