+++ sample-app/src/main/scala/example/application/projection/ConcertProjectionRepository.scala
@@ -1,6 +1,7 @@
-package example.adapter
+package example.application.projection
 
 import akka.Done
+import example.adapter.ConcertEvent
 import slick.basic.DatabaseConfig
 import slick.dbio.DBIO
 import slick.jdbc.JdbcProfile
