+++ sample-app/src/main/scala/example/application/projection/DefaultConcertProjectionRepository.scala
@@ -1,7 +1,7 @@
 package example.application.projection
 
 import akka.Done
-import example.adapter.{ ConcertEvent, ConcertProjectionRepository }
+import example.adapter.ConcertEvent
 import example.adapter.ConcertEvent.{ ConcertCancelled, ConcertCreated, ConcertTicketsBought }
 import example.application.ApplicationExecutionContext
 import example.readmodel.ConcertDatabaseService
